Pressure plates and doors added
There's some issues with the animations and placing the button where we
want. I found a work around with this by parenting a new gameObject with
the pressureplate gameobject. Otherwise the pressureplate would just
return back to where the animation starts.

We can talk about this in person in our next meeting. Will get onto
death conditions and tinker with this some more later. :)