Added basic graphic things and polished pressure plates
Because Unity is kind of bad at local animation changes, I had to parent
the pressure plates with another game object, which is animated in place
of the pressure plates. Dragged them in as prefabs, but there's a trick
to using them.

Will tell you guys during our next meeting.