Energy and health bars added
Also made the regen energy work. I added a penalty for the player if
their energy ever reaches 0. Takes a bit before it starts regening if it
reaches that point, other it regens pretty quickly.

Added an "energyCost" variable to the weapon super class, this way we
can easily say how much each weapon costs in terms of energy
consumption.