Modified bullets and added changeable equipment/attachable models to a limited extent
Having some issues working out how to store a string path in each
weapon: club, pistol, etc.

The problem is that you need the mesh path to instantiate the weapon,
but you can only assign that path to a string once it's instantiated.

Local variables in Club or Pistol don't seem to work either, since we're
storing the weapon as a "Weapon" reference, accessing members in any
class inheriting from this doesn't seems to work, unless they were
already declared in the super and assigned later.