OverWorld
Have got the camera and boject moving correctly around the map.

The Colliosion does not work when it is just the object and the terrian
for some reason. So to fix this I have implemented 3 circles and flatted
them down to look like sand around the islands. The object now collides
with the circles so it is a work around.