Fixed some bugs, added new animations, updated club to pipe
Had to make changes in weapon class. It's better to include the collider
in a seperate gameObject to the actual weapon or the weapon mesh, so
that it can be rotated/move freely to be in the best position. Because
of this, col.gameObject was changed to
col.transform.FindChild("collider").gameObject :B